# backend/api/v1/handlers.go
package v1

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/jayden/personal-blog-backend/models"
)

// 响应结构体
// @Description 统一API响应格式
type Response struct {
	// 状态码
	Code int `json:"code" example:"200"`
	// 数据
	Data interface{} `json:"data"`
	// 消息
	Msg string `json:"msg" example:"success"`
	// 跟踪ID
	TraceId string `json:"trace_id" example:"1234567890"`
}

// 分页响应结构体
// @Description 分页响应格式
type PageResponse struct {
	// 页码
	Page int `json:"page" example:"1"`
	// 每页数量
	PageSize int `json:"page_size" example:"10"`
	// 总数
	Total int64 `json:"total" example:"100"`
	// 数据列表
	List interface{} `json:"list"`
}

// @Summary 获取相册列表
// @Description 获取相册列表，支持分页
// @Tags 相册
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取相册列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/album/find_album_list [post]
func FindAlbumListHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取相册列表逻辑
	response := Response{
		Code: 200,
		Data: []models.Album{},
		Msg:  "获取相册列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取相册下的照片列表
// @Description 根据相册ID获取照片列表，支持分页
// @Tags 相册
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取照片列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/album/find_photo_list [post]
func FindPhotoListHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取相册下的照片列表逻辑
	response := Response{
		Code: 200,
		Data: []models.Photo{},
		Msg:  "获取照片列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取相册详情
// @Description 根据相册ID获取相册详情
// @Tags 相册
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取相册成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/album/get_album [post]
func GetAlbumHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取相册详情逻辑
	response := Response{
		Code: 200,
		Data: models.Album{},
		Msg:  "获取相册成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 文章归档
// @Description 获取文章归档(时间轴)
// @Tags 文章
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取文章归档成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/article/get_article_archives [post]
func GetArticleArchivesHandler(w http.ResponseWriter, r *http.Request) {
	// 实现文章归档逻辑
	response := Response{
		Code: 200,
		Data: []map[string]interface{}{},
		Msg:  "获取文章归档成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 通过分类获取文章列表
// @Description 通过分类获取文章列表，支持分页
// @Tags 文章
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取分类文章列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/article/get_article_classify_category [post]
func GetArticleClassifyCategoryHandler(w http.ResponseWriter, r *http.Request) {
	// 实现通过分类获取文章列表逻辑
	response := Response{
		Code: 200,
		Data: []models.Article{},
		Msg:  "获取分类文章列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 通过标签获取文章列表
// @Description 通过标签获取文章列表，支持分页
// @Tags 文章
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取标签文章列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/article/get_article_classify_tag [post]
func GetArticleClassifyTagHandler(w http.ResponseWriter, r *http.Request) {
	// 实现通过标签获取文章列表逻辑
	response := Response{
		Code: 200,
		Data: []models.Article{},
		Msg:  "获取标签文章列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取文章详情
// @Description 根据文章ID获取文章详情
// @Tags 文章
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取文章详情成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/article/get_article_details [post]
func GetArticleDetailsHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取文章详情逻辑
	response := Response{
		Code: 200,
		Data: map[string]interface{}{},
		Msg:  "获取文章详情成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取首页文章列表
// @Description 获取首页文章列表，支持分页
// @Tags 文章
// @Accept  json
// @Produce  json
// @Param page query int false "页码" default(1)
// @Param limit query int false "每页数量" default(10)
// @Success 200 {object} Response "获取首页文章列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/article/get_article_home_list [post]
func GetArticleHomeListHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取首页文章列表逻辑
	// 获取分页参数
	pageStr := r.URL.Query().Get("page")
	limitStr := r.URL.Query().Get("limit")
	page := 1
	limit := 10

	if pageStr != "" {
		if p, err := strconv.Atoi(pageStr); err == nil && p > 0 {
			page = p
		}
	}

	if limitStr != "" {
		if l, err := strconv.Atoi(limitStr); err == nil && l > 0 {
			limit = l
		}
	}

	offset := (page - 1) * limit

	// 获取文章列表
	articles, err := models.GetArticles(limit, offset)
	if err != nil {
		response := Response{
			Code: 500,
			Data: nil,
			Msg:  "获取文章列表失败: " + err.Error(),
		}
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(response)
		return
	}

	// 获取文章总数
	// 暂时使用固定值，后续需要实现获取文章总数的方法
	total := int64(0)

	pageResponse := PageResponse{
		Page:     page,
		PageSize: limit,
		Total:    total,
		List:     articles,
	}

	response := Response{
		Code: 200,
		Data: pageResponse,
		Msg:  "获取首页文章列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取首页推荐文章列表
// @Description 获取首页推荐文章列表
// @Tags 文章
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取推荐文章列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/article/get_article_recommend [post]
func GetArticleRecommendHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取首页推荐文章列表逻辑
	response := Response{
		Code: 200,
		Data: PageResponse{
			List: []models.Article{},
		},
		Msg:  "获取推荐文章列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 点赞文章
// @Description 点赞文章
// @Tags 文章
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "点赞文章成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/article/like_article [post]
func LikeArticleHandler(w http.ResponseWriter, r *http.Request) {
	// 实现点赞文章逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "点赞文章成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 查询评论列表
// @Description 查询评论列表，支持分页
// @Tags 评论
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取评论列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/comment/find_comment_list [post]
func FindCommentListHandler(w http.ResponseWriter, r *http.Request) {
	// 实现查询评论列表逻辑
	response := Response{
		Code: 200,
		Data: []models.Comment{},
		Msg:  "获取评论列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 查询最新评论回复列表
// @Description 查询最新评论回复列表
// @Tags 评论
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取最新评论列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/comment/find_comment_recent_list [post]
func FindCommentRecentListHandler(w http.ResponseWriter, r *http.Request) {
	// 实现查询最新评论回复列表逻辑
	response := Response{
		Code: 200,
		Data: []models.Comment{},
		Msg:  "获取最新评论列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 查询评论回复列表
// @Description 查询评论回复列表，支持分页
// @Tags 评论
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取评论回复列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/comment/find_comment_reply_list [post]
func FindCommentReplyListHandler(w http.ResponseWriter, r *http.Request) {
	// 实现查询评论回复列表逻辑
	response := Response{
		Code: 200,
		Data: []models.Comment{},
		Msg:  "获取评论回复列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 创建评论
// @Description 创建评论
// @Tags 评论
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "创建评论成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/comment/add_comment [post]
func AddCommentHandler(w http.ResponseWriter, r *http.Request) {
	// 实现创建评论逻辑
	response := Response{
		Code: 200,
		Data: models.Comment{},
		Msg:  "创建评论成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 点赞评论
// @Description 点赞评论
// @Tags 评论
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "点赞评论成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/comment/like_comment [post]
func LikeCommentHandler(w http.ResponseWriter, r *http.Request) {
	// 实现点赞评论逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "点赞评论成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 更新评论
// @Description 更新评论
// @Tags 评论
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "更新评论成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/comment/update_comment [post]
func UpdateCommentHandler(w http.ResponseWriter, r *http.Request) {
	// 实现更新评论逻辑
	response := Response{
		Code: 200,
		Data: models.Comment{},
		Msg:  "更新评论成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取博客前台首页信息
// @Description 获取博客前台首页信息
// @Tags 博客
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取博客前台首页信息成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/blog [get]
func GetBlogHomeInfoHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取博客前台首页信息逻辑
	response := Response{
		Code: 200,
		Data: map[string]interface{}{
			"article_count":         0,
			"category_count":        0,
			"tag_count":             0,
			"total_user_view_count": 0,
			"total_page_view_count": 0,
			"page_list":             []interface{}{},
			"website_config": map[string]interface{}{
				"admin_url":      "",
				"websocket_url":  "",
				"tourist_avatar": "",
				"user_avatar":    "",
				"website_feature": map[string]interface{}{
					"is_chat_room":      0,
					"is_comment_review": 0,
					"is_email_notice":   0,
					"is_message_review": 0,
					"is_music_player":   0,
					"is_reward":         0,
				},
				"social_login_list": []interface{}{},
				"social_url_list":   []interface{}{},
			},
		},
		Msg: "获取博客前台首页信息成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 删除用户绑定第三方平台账号
// @Description 删除用户绑定第三方平台账号
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "删除用户绑定第三方平台账号成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/delete_user_bind_third_party [post]
func DeleteUserBindThirdPartyHandler(w http.ResponseWriter, r *http.Request) {
	// 实现删除用户绑定第三方平台账号逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "删除用户绑定第三方平台账号成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取用户信息
// @Description 获取用户信息
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取用户信息成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/get_user_info [get]
func GetUserInfoHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取用户信息逻辑
	response := Response{
		Code: 200,
		Data: models.UserInfo{},
		Msg:  "获取用户信息成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取用户点赞列表
// @Description 获取用户点赞列表
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取用户点赞列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/get_user_like [get]
func GetUserLikeHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取用户点赞列表逻辑
	response := Response{
		Code: 200,
		Data: models.UserLike{},
		Msg:  "获取用户点赞列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 修改用户头像
// @Description 修改用户头像
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "修改用户头像成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/update_user_avatar [post]
func UpdateUserAvatarHandler(w http.ResponseWriter, r *http.Request) {
	// 实现修改用户头像逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "修改用户头像成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 修改用户绑定邮箱
// @Description 修改用户绑定邮箱
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "修改用户绑定邮箱成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/update_user_bind_email [post]
func UpdateUserBindEmailHandler(w http.ResponseWriter, r *http.Request) {
	// 实现修改用户绑定邮箱逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "修改用户绑定邮箱成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 修改用户绑定手机号
// @Description 修改用户绑定手机号
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "修改用户绑定手机号成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/update_user_bind_phone [post]
func UpdateUserBindPhoneHandler(w http.ResponseWriter, r *http.Request) {
	// 实现修改用户绑定手机号逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "修改用户绑定手机号成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 修改用户绑定第三方平台账号
// @Description 修改用户绑定第三方平台账号
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "修改用户绑定第三方平台账号成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/update_user_bind_third_party [post]
func UpdateUserBindThirdPartyHandler(w http.ResponseWriter, r *http.Request) {
	// 实现修改用户绑定第三方平台账号逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "修改用户绑定第三方平台账号成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 修改用户信息
// @Description 修改用户信息
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "修改用户信息成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/update_user_info [post]
func UpdateUserInfoHandler(w http.ResponseWriter, r *http.Request) {
	// 实现修改用户信息逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "修改用户信息成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取关于我的信息
// @Description 获取关于我的信息
// @Tags 博客
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取关于我的信息成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/blog/about_me [get]
func GetAboutMeHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取关于我的信息逻辑
	response := Response{
		Code: 200,
		Data: map[string]interface{}{
			"content": "关于我的信息内容",
		},
		Msg: "获取关于我的信息成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 修改用户密码
// @Description 修改用户密码
// @Tags 用户
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "修改用户密码成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/user/update_user_password [post]
func UpdateUserPasswordHandler(w http.ResponseWriter, r *http.Request) {
	// 实现修改用户密码逻辑
	response := Response{
		Code: 200,
		Data: nil,
		Msg:  "修改用户密码成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取游客信息
// @Description 获取游客信息
// @Tags 游客
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取游客信息成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/get_tourist_info [get]
func GetTouristInfoHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取游客信息逻辑
	response := Response{
		Code: 200,
		Data: map[string]interface{}{
			"nickname": "游客",
			"avatar":   "",
		},
		Msg: "获取游客信息成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// @Summary 获取说说列表
// @Description 获取说说列表，支持分页
// @Tags 说说
// @Accept  json
// @Produce  json
// @Success 200 {object} Response "获取说说列表成功"
// @Failure 500 {object} Response "服务器错误"
// @Router /blog-api/v1/talk/find_talk_list [post]
func FindTalkListHandler(w http.ResponseWriter, r *http.Request) {
	// 实现获取说说列表逻辑
	response := Response{
		Code: 200,
		Data: PageResponse{
			Page:     1,
			PageSize: 10,
			Total:    0,
			List:     []interface{}{},
		},
		Msg: "获取说说列表成功",
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}