# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  v1.Response:
    description: 统一API响应格式
    properties:
      code:
        description: 状态码
        example: 200
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        example: success
        type: string
      trace_id:
        description: 跟踪ID
        example: "1234567890"
        type: string
    type: object
host: localhost:8083
info:
  contact: {}
  description: 这是一个博客系统的 API 文档
  title: 博客系统 API接口文档
  version: "1.0"
paths:
...
  /blog-api/v1/album/find_album_list:
    post:
      consumes:
      - application/json
      description: 获取相册列表，支持分页
      produces:
      - application/json
      responses:
        "200":
          description: 获取相册列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取相册列表
      tags:
      - 相册
  /blog-api/v1/album/find_photo_list:
    post:
      consumes:
      - application/json
      description: 根据相册ID获取照片列表，支持分页
      produces:
      - application/json
      responses:
        "200":
          description: 获取照片列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取相册下的照片列表
      tags:
      - 相册
  /blog-api/v1/album/get_album:
    post:
      consumes:
      - application/json
      description: 根据相册ID获取相册详情
      produces:
      - application/json
      responses:
        "200":
          description: 获取相册成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取相册详情
      tags:
      - 相册
  /blog-api/v1/article/get_article_archives:
    post:
      consumes:
      - application/json
      description: 获取文章归档(时间轴)
      produces:
      - application/json
      responses:
        "200":
          description: 获取文章归档成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 文章归档
      tags:
      - 文章
  /blog-api/v1/article/get_article_classify_category:
    post:
      consumes:
      - application/json
      description: 通过分类获取文章列表，支持分页
      produces:
      - application/json
      responses:
        "200":
          description: 获取分类文章列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 通过分类获取文章列表
      tags:
      - 文章
  /blog-api/v1/article/get_article_classify_tag:
    post:
      consumes:
      - application/json
      description: 通过标签获取文章列表，支持分页
      produces:
      - application/json
      responses:
        "200":
          description: 获取标签文章列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 通过标签获取文章列表
      tags:
      - 文章
  /blog-api/v1/article/get_article_details:
    post:
      consumes:
      - application/json
      description: 根据文章ID获取文章详情
      produces:
      - application/json
      responses:
        "200":
          description: 获取文章详情成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取文章详情
      tags:
      - 文章
  /blog-api/v1/article/get_article_home_list:
    post:
      consumes:
      - application/json
      description: 获取首页文章列表，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取首页文章列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取首页文章列表
      tags:
      - 文章
  /blog-api/v1/article/get_article_recommend:
    post:
      consumes:
      - application/json
      description: 获取首页推荐文章列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取推荐文章列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取首页推荐文章列表
      tags:
      - 文章
  /blog-api/v1/article/like_article:
    post:
      consumes:
      - application/json
      description: 点赞文章
      produces:
      - application/json
      responses:
        "200":
          description: 点赞文章成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 点赞文章
      tags:
      - 文章
  /blog-api/v1/comment/add_comment:
    post:
      consumes:
      - application/json
      description: 创建评论
      produces:
      - application/json
      responses:
        "200":
          description: 创建评论成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 创建评论
      tags:
      - 评论
  /blog-api/v1/comment/find_comment_list:
    post:
      consumes:
      - application/json
      description: 查询评论列表，支持分页
      produces:
      - application/json
      responses:
        "200":
          description: 获取评论列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询评论列表
      tags:
      - 评论
  /blog-api/v1/comment/find_comment_recent_list:
    post:
      consumes:
      - application/json
      description: 查询最新评论回复列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取最新评论列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询最新评论回复列表
      tags:
      - 评论
  /blog-api/v1/comment/find_comment_reply_list:
    post:
      consumes:
      - application/json
      description: 查询评论回复列表，支持分页
      produces:
      - application/json
      responses:
        "200":
          description: 获取评论回复列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询评论回复列表
      tags:
      - 评论
  /blog-api/v1/comment/like_comment:
    post:
      consumes:
      - application/json
      description: 点赞评论
      produces:
      - application/json
      responses:
        "200":
          description: 点赞评论成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 点赞评论
      tags:
      - 评论
  /blog-api/v1/comment/update_comment:
    post:
      consumes:
      - application/json
      description: 更新评论
      produces:
      - application/json
      responses:
        "200":
          description: 更新评论成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 更新评论
      tags:
      - 评论
  /blog-api/v1/user/delete_user_bind_third_party:
    post:
      consumes:
      - application/json
      description: 删除用户绑定第三方平台账号
      produces:
      - application/json
      responses:
        "200":
          description: 删除用户绑定第三方平台账号成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 删除用户绑定第三方平台账号
      tags:
      - 用户
  /blog-api/v1/user/get_user_info:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户信息成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取用户信息
      tags:
      - 用户
  /blog-api/v1/user/get_user_like:
    get:
      consumes:
      - application/json
      description: 获取用户点赞列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户点赞列表成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取用户点赞列表
      tags:
      - 用户
  /blog-api/v1/user/update_user_avatar:
    post:
      consumes:
      - application/json
      description: 修改用户头像
      produces:
      - application/json
      responses:
        "200":
          description: 修改用户头像成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改用户头像
      tags:
      - 用户
  /blog-api/v1/user/update_user_bind_email:
    post:
      consumes:
      - application/json
      description: 修改用户绑定邮箱
      produces:
      - application/json
      responses:
        "200":
          description: 修改用户绑定邮箱成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改用户绑定邮箱
      tags:
      - 用户
  /blog-api/v1/user/update_user_bind_phone:
    post:
      consumes:
      - application/json
      description: 修改用户绑定手机号
      produces:
      - application/json
      responses:
        "200":
          description: 修改用户绑定手机号成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改用户绑定手机号
      tags:
      - 用户
  /blog-api/v1/user/update_user_bind_third_party:
    post:
      consumes:
      - application/json
      description: 修改用户绑定第三方平台账号
      produces:
      - application/json
      responses:
        "200":
          description: 修改用户绑定第三方平台账号成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改用户绑定第三方平台账号
      tags:
      - 用户
  /blog-api/v1/user/update_user_info:
    post:
      consumes:
      - application/json
      description: 修改用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 修改用户信息成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改用户信息
      tags:
      - 用户
  /blog-api/v1/user/update_user_password:
    post:
      consumes:
      - application/json
      description: 修改用户密码
      produces:
      - application/json
      responses:
        "200":
          description: 修改用户密码成功
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改用户密码
      tags:
      - 用户
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"